Expand SQL formatter to LakeFormation
