+++ awswrangler/_sql_formatter.py
@@ -1,8 +1,9 @@
-"""Formatting logic for Athena parameters."""
+"""Formatting logic for SQL parameters."""
 import datetime
 import decimal
+import re
 from enum import Enum
-from typing import Any, Dict, Generic, Sequence, Type, TypeVar
+from typing import Any, Dict, Generic, Optional, Sequence, Type, TypeVar
 
 
 class _EngineType(Enum):
@@ -165,3 +166,26 @@ def _format_parameters(params: Dict[str, Any], engine: _EngineType) -> Dict[str,
         processed_params[k] = str(abs_type)
 
     return processed_params
+
+
+_PATTERN = re.compile(r":([A-Za-z0-9_]+)(?![A-Za-z0-9_])")
+
+
+def _process_sql_params(sql: str, params: Optional[Dict[str, Any]], engine: _EngineType = _EngineType.PRESTO) -> str:
+    if params is None:
+        params = {}
+
+    processed_params = _format_parameters(params, engine=engine)
+
+    def replace(match: re.Match) -> str:  # type: ignore
+        key = match.group(1)
+
+        if key not in processed_params:
+            # do not replace anything if the parameter is not provided
+            return str(match.group(0))
+
+        return str(processed_params[key])
+
+    sql = _PATTERN.sub(replace, sql)
+
+    return sql
